Move mob flow kernel wrappers into runtime ownership
