+++ meerkat-mob/src/lib.rs
@@ -39,7 +39,6 @@ mod build;
 pub mod definition;
 pub mod error;
 pub mod event;
-pub mod generated;
 pub mod ids;
 pub mod launch;
 #[doc(hidden)]
@@ -87,6 +86,7 @@ pub use runtime::bridge_protocol::{
     BridgeMemberRuntimeState, BridgeObservationResponse, BridgePeerConnectivity, BridgePeerSpec,
     BridgePeerWiringPayload, BridgeReply, BridgeRetireResponse, BridgeSupervisorPayload,
 };
+pub use runtime::flow_kernels;
 #[cfg(feature = "runtime-adapter")]
 pub use runtime::local_bridge::LocalMobRuntimeBridge;
 pub use runtime::{FlowFrameKernel, FlowFrameMutator};
