+++ meerkat-mob/src/runtime/flow_frame_engine.rs
@@ -9,15 +9,15 @@
 
 use crate::definition::{FlowNodeSpec, FrameSpec};
 use crate::error::MobError;
-use crate::generated::flow_frame_loop_driver::{
-    FlowFrameLoopDecision, FlowFrameLoopDriver, FlowFrameLoopStorePlan, FlowFrameLoopWork,
-    FlowFrameTerminalPhase,
-};
-use crate::generated::flow_run;
 use crate::ids::{FlowNodeId, FrameId, LoopId, LoopInstanceId, RunId, StepId};
 use crate::run::{FlowContext, FrameSnapshot, LoopSnapshot, MobRun, StepRunStatus};
 use crate::runtime::conditions::evaluate_condition;
 use crate::runtime::flow_frame_kernel::{FlowFrameKernel, FlowFrameMutator, StepCompletionOpts};
+use crate::runtime::flow_frame_loop_driver::{
+    FlowFrameLoopDecision, FlowFrameLoopDriver, FlowFrameLoopStorePlan, FlowFrameLoopWork,
+    FlowFrameTerminalPhase,
+};
+use crate::runtime::flow_kernels::flow_run;
 use crate::store::MobRunStore;
 use async_trait::async_trait;
 use futures::stream::{FuturesUnordered, StreamExt};
@@ -987,7 +987,7 @@ impl FlowFrameEngine {
         root_frame_id: &FrameId,
         root_spec: &FrameSpec,
         context: &FlowContext,
-        obligation: crate::generated::protocol_flow_loop_until_evaluation::FlowLoopUntilEvaluationObligation,
+        obligation: crate::runtime::flow_loop_until_evaluation::FlowLoopUntilEvaluationObligation,
     ) -> Result<(), MobError> {
         for _ in 0..=5usize {
             let run = self.require_run(run_id).await?;
