+++ meerkat-mob/src/runtime/flow_kernels/flow_frame.rs
@@ -1,4 +1,4 @@
-// Generated by `cargo xtask machine-codegen --all`.
+// Local flow-frame kernel wrapper derived from the former compat codegen.
 #![allow(dead_code)]
 
 use meerkat_machine_kernels::{
