+++ meerkat-mob/src/runtime/mod.rs
@@ -77,6 +77,9 @@ mod events;
 mod flow;
 pub mod flow_frame_engine;
 pub mod flow_frame_kernel;
+mod flow_frame_loop_driver;
+pub mod flow_kernels;
+mod flow_loop_until_evaluation;
 mod flow_run_kernel;
 mod handle;
 #[cfg(feature = "runtime-adapter")]
