+++ meerkat-mob/tests/flow_frame_kernel.rs
@@ -7,7 +7,7 @@
 )]
 
 use meerkat_machine_kernels::{KernelInput, KernelState, KernelValue};
-use meerkat_mob::generated::flow_frame;
+use meerkat_mob::runtime::flow_kernels::flow_frame;
 use std::collections::{BTreeMap, BTreeSet};
 
 // Helper: build a KernelValue::NamedVariant for NodeRunStatus
