+++ meerkat-mob/tests/loop_iteration_kernel_tests.rs
@@ -1,7 +1,7 @@
 #![allow(clippy::expect_used, clippy::unwrap_used)]
 
 use meerkat_machine_kernels::{KernelInput, KernelValue};
-use meerkat_mob::generated::loop_iteration;
+use meerkat_mob::runtime::flow_kernels::loop_iteration;
 use std::collections::BTreeMap;
 
 fn str_val(s: &str) -> KernelValue {
